Feat: delete old docs and replace them with new ones on command
